refactor: signUpBenefits component
