+++ sites/public/src/pages/sign-in.tsx
@@ -14,8 +14,7 @@ import {
 } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../lib/constants"
 import { EnumUserErrorExtraModelUserErrorMessages } from "@bloom-housing/backend-core/types"
-import { faStopwatch, faEye, faLock } from "@fortawesome/free-solid-svg-icons"
-import { Icon, HeadingGroup } from "@bloom-housing/ui-seeds"
+import SignUpBenefits from "../components/account/SignUpBenefits"
 
 const SignIn = () => {
   const { login, userService } = useContext(AuthContext)
@@ -127,11 +126,6 @@ const SignIn = () => {
     }
   }, [networkError])
 
-  const iconListItems = [
-    { icon: faStopwatch, text: "Apply faster with saved application details" },
-    { icon: faEye, text: "Check on the status of an application at any time" },
-    { icon: faLock, text: "Simply reset your password if you forget it" },
-  ]
   return (
     <>
       <FormsLayout>
@@ -150,27 +144,7 @@ const SignIn = () => {
             }}
             showRegisterBtn={true}
           />
-          <div className="w-min mt-6 ml-6">
-            <HeadingGroup
-              heading={"Sign up quickly and check application status at anytime"}
-              subheading={
-                "Having an account will save you time by using saved application details, and allow you to check the status of an application at anytime."
-              }
-              size="xl"
-            />
-            <ul className="flex flex-col w-min">
-              {iconListItems.map((item) => (
-                <li className="flex flex-row w-max mb-2 items-center">
-                  <Icon
-                    icon={item.icon}
-                    size="lg"
-                    className="border border-white bg-white rounded-full p-1.5 w-max"
-                  />
-                  <p className="ml-2">{item.text}</p>
-                </li>
-              ))}
-            </ul>
-          </div>
+          <SignUpBenefits />
         </div>
       </FormsLayout>
